fix: avoid infinite predictions (#3)
The BetaGeoFitter's method conditional_expected_number_of_purchases_up_to_time is numerically unstable because of the scipy.special.hyp2f1 function used in it:

When frequency=0 and recency=0 and the prediction horizon t is above the customer age T by a sufficient amount, the hyp2f1 function produces negative numbers. This results in RuntimeWarning when trying to take the log from them and leads to NaN predictions.
When the frequency is very high, the model runs into overflow errors (when a number is too large to store as float64), leading to the inf output of hyp2f1 and inf prediction as the result.
The master version at the time of this PR is handling okay the 2) issue but fails with the 1) one.

Apparently, there were a few attempts (PR and PR) at fixing the case 1). However, by using them I still ran into errors again in case 2) with high-frequency users. Moreover, the mpmath.hyp2f1 function does not get broadcasted well with numpy and the predict process takes ages.

This PR takes the best from the two worlds and uses one of the approaches above, depending on the case.